ai_chat_moderation/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
python-dotenv==1.1.0
Requests==2.32.3
